start working on middleware
