--- src/app/ClientMember/page.jsx
@@ -1,23 +1,0 @@
-'use client'
-
-import { useSession } from 'next-auth/react'
-import { redirect } from 'next/navigation'
-
-const Member = async () => {
-  const { data: session } = useSession({
-    required: true,
-    onUnauthenticated() {
-      redirect('/api/auth/signin?callbackUrl=/ClientMember')
-    }
-  })
-
-  return (
-    <div className='pt-24'>
-      <h1>Member Client Session</h1>
-      <p>{session?.user?.email}</p>
-      <p>{session?.user?.role}</p>
-    </div>
-  )
-}
-
-export default Member
