--- src/app/Denied/page.jsx
@@ -1,8 +1,0 @@
-const Denied = () => {
-  return (
-    <div>
-      <h1 className="pt-24">Denied</h1>
-    </div>
-  );
-}
-export default Denied;
