--- src/app/Member/page.jsx
@@ -1,21 +1,0 @@
-import { getServerSession } from 'next-auth'
-import { options } from '../api/(auth)/auth/[...nextauth]/options'
-import { redirect } from 'next/navigation'
-
-const Member = async () => {
-
-  const session = await getServerSession(options)
-
-  if(!session) {
-    redirect('/api/auth/signin?callbackUrl=/Member')
-  }
-
-  return (
-    <div className='pt-24'>
-      <h1>Member Server Session</h1>
-      <p>{session?.user?.email}</p>
-      <p>{session?.user?.role}</p>
-    </div>
-  );
-}
-export default Member;
