--- src/app/Public/page.jsx
@@ -1,6 +1,0 @@
-const PublicRoute = () => {
-  return (
-    <p className="pt-24">PublicRoute</p>
-  );
-}
-export default PublicRoute;
