+++ src/app/account/[...id]/page.jsx
@@ -1,0 +1,11 @@
+'use client'
+
+const AccountIndividualPage = ({params}) => {
+  console.log(params)
+  return (
+    <p className="pt-24">
+      Account individual
+    </p>
+  );
+}
+export default AccountIndividualPage;
